refactor(cli): move the lead renderer under scripts/ with a declaration mirror
Two facts forced the move, both measured rather than guessed:

  - a .ts test importing the untyped .mjs is TS7016 ("Could not find a
    declaration file"), and packages/cli's hidden test layer is ledgered in
    check-type-check-coverage.mjs with "the first new error in it should go red
    rather than be absorbed". The fix is the hand-written .d.mts this repo
    already uses three times over -- and check-declaration-mirrors only
    discovers scripts/**/*.d.mts, so a declaration under packages/cli/bin would
    have been exactly the unwatched drift that gate exists to prevent.
  - that gate import()s the module with bare node, so the module may not import
    a .ts. The CLI's name therefore arrives as a parameter; the shim already
    imports invocation.ts on its failure path and owns that coupling.

Measured after the move: the two new test files add 0 errors to the layer, with
a two-sided control (the same program reports TS7016 for the undeclared sibling
import on the line above and nothing for this one).
